+++ include/admin_events.inc.php
@@ -124,3 +124,81 @@ function facial_loc_end_element_set_unit()
     ));
     $template->append('PLUGINS_BATCH_MANAGER_UNIT_ELEMENT_SUBTEMPLATE', 'plugins/facial/template/batch_manager_unit.tpl');
 }
+
+function facial_batch_global()
+{
+  global $template, $logger;
+
+  $logger->debug('Adding facial options to batch manager');
+
+  load_language('plugin.lang', FACIAL_PATH);
+
+  // Assign the template for batch management
+  $template->set_filename('FACIAL_batch_global', dirname(__FILE__).'./template/batch_global.tpl');
+
+  $FacialOptions = array();
+  $template->assign('FacialOptions', $FacialOptions);
+
+
+  // Add info on the "choose action" dropdown in the batch manager
+  $template->append('element_set_global_plugins_actions', array(
+    'ID' => 'facial', // ID of the batch manager action
+    'NAME' => l10n('Recognize Faces'), // Description of the batch manager action
+    'CONTENT' => ''
+  ));
+}
+
+// Process the submit action
+function facial_batch_global_submit($action, $collection)
+{
+  global $logger;
+  $logger->debug('facial_batch_global_submit called with action: '.$action);
+
+  if ($action == 'facial')
+  {
+    // Process facial recognition on the selected images
+    foreach ($collection as $image_id)
+    {
+      $logger->debug("Processing image ID: $image_id");
+      //facial_process_image($image_id);
+
+      /*
+        For each image in the collection,
+        I want to first send the image to CompreFace
+        facial recognition service. It will return an
+        array of recognized faces. Then i want to 'tag'
+        the image with who is in the image.
+
+        If a face isn't recognized, I want to add an
+        unknown subject to CompreFace and tag the image with
+        'unknown-face' tag.
+
+        I consider a face with a similarity score of less than .95 as
+        an unknown face.
+      */
+      $faces = facial_recognize_faces_by_image_id($image_id);
+      foreach ($faces as $face)
+      {
+        $logger->debug('Face detected with subjects: '.var_export($face['subjects'], true));
+        if (count($face['subjects']) > 0 && $face['subjects'][0]['similarity'] < 0.95)
+        {
+          // unknown face, it has a subject, but under the threshold of similarity that we know about
+          $logger->debug("Unknown face: " . count($face['subjects']) . " subjects, similarity: " . $face['subjects'][0]['similarity']);
+          $logger->debug('Unknown face detected, adding unknown-face tag');
+          facial_add_tag_to_image('unknown-face', $image_id);
+
+          // If it's an unknown face, I'd like to generate a UUID and add it to the CompreFace system
+          // What will need to happen though is we'll need a system to "combine" known faces together
+        }
+        elseif (count($face['subjects']) > 0 && $face['subjects'][0]['similarity'] >= 0.95)
+        {
+          // recognized face
+          $subject_name = $face['subjects'][0]['subject'];
+          $logger->debug("Recognized face: $subject_name, adding tag to image");
+          facial_add_tag_to_image($subject_name, $image_id);
+        }
+      }
+    }
+  }
+  return;
+}